fix: address sourcery-ai review + split tests by concern
- Use executor.submit() instead of per-event daemon threads
- Include template string in UPSRSValidationError message
- Test callback lock released on exception
- Fix CLI success-path tests to assert exact exit code
- Split test_bug_fixes.py into test_thread_safety.py,
  test_websocket_config.py, test_cli_validation.py
